Fix CheckCallback return type

+++ include/psxsdk/libetc.h
@@ -11,7 +11,7 @@ struct Callbacks {
     int (*RestartCallback)(void);
 };
 
-extern u16 D_8002C2BA;
+extern unsigned short D_8002C2BA;
 extern struct Callbacks* D_8002D340;
 
 int VSync(int mode);
@@ -22,7 +22,7 @@ void* DMACallback(int dma, void (*func)());
 int ResetCallback(void);
 int StopCallback(void);
 int RestartCallback(void);
-u16 CheckCallback(void);
+int CheckCallback(void);
 
 long SetVideoMode(long mode);
